Extract model factory compat helper
